Parser file refactor
I did this to make it easier to work with the source code. I decided
to do this because editing parser.cpp was lagging Vim.
